+++ spec/document/sdt_in_document_spec.rb
@@ -5,4 +5,9 @@ describe 'Sdt in document' do
     docx = OoxmlParser::DocxParser.parse_docx('spec/document/sdt/sdt_in_document.docx')
     expect(docx.elements.first.sdt_content.paragraphs.first.runs.first.text).to eq('Simple Test Text')
   end
+
+  it 'sdt_with_table.docx' do
+    docx = OoxmlParser::DocxParser.parse_docx('spec/document/sdt/sdt_with_table.docx')
+    expect(docx.elements.first.sdt_content.elements.first).to be_a(OoxmlParser::Table)
+  end
 end